Reorganise and improve generation of hex files for various atmega.
You can now do

  AVR_FREQ=16000000L BAUD_RATE=57600 make atmega328 atmega168 atmega88 atmega8

to create hex files with more useful names all at once, the file names
generated include the frequency and baud rate.

For "safety" the defaults, including the fuse settings, have been
reduced to 8MHz internal oscillator, this reduces the chance of you
doing a make xxx_isp and having it burn fuses to brick (all be it
recoverably) your AVR that doesn't have a crystal.

A lot of descriptive comment has been added to the new Makefile.generic
which includes the specifications for the above chips now.