Arduino ISP  'speed' and 'protocol'  setting
 I have tested with Arduino uno R3 as ISP to program a ATmega48pa mcu, which resulted in the following error : 
"stk500_getsync() attempt 2 of 10: not in sync: resp=0x1c"

The error does not occurs while uploading the bootloader, but only when uploading sketches. 
So i noticed the difference between  commands and singled out the problem causing settings(speed,protocol). 

However i am a beginner, that is why i added the new settings as a new menu option.

tested on Arduino IDE 1.6.4