Fix makedist to select better platform.json template.
Err, actually, because version numbers can not have zero prefixed components
  https://github.com/sleemanj/optiboot/commit/7a318f3d345bfe4f4e274d9e3683fcaca34503cb
and because git doesn't store timestamps (sigh), it is hard to sort these.

But it is almost certain that you will do a couple of makedist before
being ready to go, so you can do the first one, adjust the template it made
and then the next time you makedist it will use that template to recreate itself.

Or worst case, just touch the *_platform.json you want to be the template.