Allow to ignore the baud check and compile anyway by specifying `FORCE_BAUD=1` environment variable.
NB: I was mistaken, 57600 works at 8MHz internal oscillators if you tune it, or at least it did for me, but it's still probably quite marginal.  Tuning with my example only shifted the OSCCAL by 1, if that's enough to throw it out then any fluctuation in temperature voltage etc after tuning is likely to cause problems at least with CH340G (which has a stated maximum 2% tolerance for receiving).